Update README.md
Since it's always been hard to explain why we'd want insecure SSH, I added a quote from the end that sums it up up at the beginning.  The key fact is that we don't want insecure SSH -- it just already gets used that way already and this is an easier way to make sure that it's both well known that it's insecure and also to manage the keys and make sure peeps follow best practices when using SSH in horrible ways.